requirements.txt
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
distlib==0.3.3
dnspython==2.1.0
et-xmlfile==1.1.0
filelock==3.4.0
fiona==1.10.0
h11==0.12.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.6.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.1.0
parsel==1.6.0
pipenv==2021.11.23
platformdirs==2.4.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pymongo==3.12.1
pyOpenSSL==21.0.0
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.6.1
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tldextract==3.2.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.14.0
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
w3lib==1.22.0
wsproto==1.0.0
zope.interface==5.4.0
